New features
Added the ability to search the databases directory
at a level above the application.
Added recursive connection to databases via exec
Added error output in case it was not possible to
connect to the database.
A separate glossary directory has been created as
a database editor program will be added later.
Comments have also been updated.